+++ src/components/Registration/College_Details/EmailVerify/EmailVerify.css
