+++ src/components/Registration/College_Details/RegComp/RegComp.css
