+++ src/components/Registration/login/Login.js
@@ -1,4 +1,8 @@
-import React, { useState, useEffect } from "react";
+
+
+
+
+import React, { useState,useEffect } from "react";
 import { Input } from "antd";
 import { Link, useNavigate } from "react-router-dom";
 import SimpleReactValidator from "simple-react-validator";
@@ -7,12 +11,16 @@ import { connect } from "react-redux";
 import axios from "axios";
 // import Loader from "../../Loader/Loader";
 // import setAuthToken from "../../User/setAuthToken";
+
+
 import Regback from "../../../assests/Regback.webp";
-import logo from "../../../assests/logo.png";
+import logo from "../../../assests/logo.png"
 import "./Login.scss";
+// import Navbar from '../../Navbar/Navbar'
+
 
 const Login = (props) => {
-  let navigate = useNavigate();
+let navigate = useNavigate();
 
   const validator = new SimpleReactValidator();
 
@@ -44,6 +52,7 @@ const Login = (props) => {
               setLoading(false);
               setVerifiederror(true);
               setNotFoundError(false);
+              
             } else if (res.data.error === "CA not found") {
               setNotFoundError(true);
               setVerifiederror(false);
@@ -75,98 +84,91 @@ const Login = (props) => {
   };
 
   useEffect(() => {
-    if (
-      user.email &&
-      user.password &&
-      /^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(user.email)
-    ) {
-      setActive(true);
-    } else {
+    if(user.email && user.password ){
+      setActive(true)
+    }else{
       setActive(false);
     }
-  }, [user]);
+  }, [user])
 
-  return (
-    <>
-      <div className="login-fullpage">
-        <img src={Regback} className="regbg" alt="" />
-        <Link to="/">
-          <img src={logo} className="logo" alt="" />
-        </Link>
+    return(
+        <>
+        <div className='login-fullpage'>
+        
+        <img src={Regback}  className='regbg'alt="" />
+        <img src={logo} className='logo' alt=""/>
+
+
 
         <div id="bg">
-          <div id="log_bg1">
-            <div id="box">
-              <div id="center">
-                <div id="login">Log In</div>
-                <form id="form" onSubmit={(e) => onSubmit(e)}>
-                  <input
-                    type="email"
-                    name="email"
-                    placeholder="Email Id *"
-                    id="email1"
-                    value={user.email}
-                    required
-                    onChange={(e) => onInputChange(e)}
-                  />
-                  {validator.message("email", user.email, "required|email", {
-                    className: "text-danger",
-                  })}
-                  <input
-                    type="password"
-                    name="password"
-                    placeholder="Password *"
-                    id="password1"
-                    value={user.password}
-                    required
-                    onChange={(e) => onInputChange(e)}
-                  />
+        <div id="log_bg1">
+        <div id="box">
+            <div id="center">
+              <div id="login">Log In</div>
+              <form  id="form" onSubmit={(e) => onSubmit(e)}>
+                <input
+                  type="email"
+                  name="email"
+                  placeholder="Email Id *"
+                  id="email1"
+                  value={user.email}
+                  required
+                  onChange={(e) => onInputChange(e)}
+                />
+                {validator.message("email", user.email, "required|email", {
+                  className: "text-danger",
+                })}
+                <input
+                  type="password"
+                  name="password"
+                  placeholder="Password *"
+                  id="password1"
+                  value={user.password}
+                  required
+                  onChange={(e) => onInputChange(e)}
+                />
+             
+              <Link to="/forgotpassword" id="forget">
+                Forgot Password?
+              </Link>
+             
 
-                  <Link to="/forgotpassword" id="forget">
-                    Forgot Password?
-                  </Link>
-
-                  <button
-                    type="submit"
-                    id="log"
-                    disabled={!active}
-                    style={
-                      active == true
-                        ? { background: "rgb(237,215,41)" }
-                        : { background: "rgb(204, 204, 204)" }
-                    }
-                  >
-                    Log In
-                  </button>
-                </form>
-                {error && (
-                  <div className="text-danger">Email/Password is Incorrect</div>
-                )}
-                {verifiederror && (
-                  <div className="text-danger">
-                    User is not verified. Please check registered mail.
-                  </div>
-                )}
-                {notFoundError && (
-                  <div className="text-danger">
-                    <div className="text-info">
-                      You are not registered as a CA.
-                    </div>
-                  </div>
-                )}
-                <div id="niche">
-                  <div id="bottom">Don't have an account?</div>
-                  <Link to="/registration">
-                    <div id="register"> Register Now!</div>
-                  </Link>
+              <button type="submit" id="log" disabled={!active} style={active == true ? {background: "rgb(237,215,41)"} : {background: "rgb(204, 204, 204)"}}>
+                Log In
+              </button>
+              </form>
+              {error && (
+                <div className="text-danger">Email/Password is Incorrect</div>
+              )}
+              {verifiederror && (
+                <div className="text-danger">
+                  User is not verified. Please check registered mail.
                 </div>
+              )}
+              {notFoundError && (
+                <div className="text-danger">
+                  <div className="text-info">
+                    You are not registered as a CA.
+                  </div>
+                </div>
+              )}
+              <div id="niche">
+                <div id="bottom">Don't have an account?</div>
+                <Link to="/registration">
+                  <div id="register"> Register Now!</div>
+                </Link>
               </div>
             </div>
           </div>
-        </div>
-      </div>
-    </>
-  );
-};
+          </div>
+          </div>
+          </div>
+        </>
+    );
+}
 
-export default Login;
+
+export default Login
+
+
+
